Extracts MainQueueDispatchDecorator to a separate file.
