+++ FeedModule/FeediOS/Feed UI/Composers/MainQueueDispatchDecorator.swift
@@ -1,0 +1,37 @@
+//
+//  MainQueueDispatchDecorator.swift
+//  FeediOS
+//
+//  Created by Omran Khoja on 2/22/22.
+//
+
+import Foundation
+import FeedModule
+
+final class MainQueueDispatchDecorator<T> {
+    private let decoratee: T
+    
+    init(decoratee: T) {
+        self.decoratee = decoratee
+    }
+    
+    func dispatch(completion: @escaping () -> Void) {
+        Thread.isMainThread ? completion() : DispatchQueue.main.async { completion() }
+    }
+}
+
+extension MainQueueDispatchDecorator: FeedLoader where T == FeedLoader {
+    func load(completion: @escaping (FeedLoader.Result) -> Void) {
+        decoratee.load { [weak self] result in
+            self?.dispatch { completion(result) }
+        }
+    }
+}
+
+extension MainQueueDispatchDecorator: FeedImageDataLoader where T == FeedImageDataLoader {
+    func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
+        decoratee.loadImageData(from: url) { [weak self] result in
+            self?.dispatch { completion(result) }
+        }
+    }
+}
